+++ packages/media-core/src/classifier/index.ts
@@ -1,2 +1,0 @@
-export * from "./types";
-export { classifyVideoFile } from "./classify-video-file";